Add args and be more tolerant of failures

+++ src/Core/Silk.NET.BuildTools/Generator.cs
@@ -26,43 +26,45 @@ namespace Silk.NET.BuildTools
 {
     public static class Generator
     {
-        public const bool TestMode = false;
-        public static void Run(Config config)
+        public static void Run(Config config, IReadOnlyList<string> args)
         {
             var tasks = new Task[config.Tasks.Length];
             var dirsToCheckThrough = config.Tasks.Select(x => x.OutputOpts.Folder).Distinct();
-            foreach (var s in dirsToCheckThrough)
+            if (!args.Contains("no-clear"))
             {
-                foreach (var file in Directory.GetFiles(s, "*.gen.cs", SearchOption.AllDirectories))
+                foreach (var s in dirsToCheckThrough)
                 {
-                    File.Delete(file);
+                    foreach (var file in Directory.GetFiles(s, "*.gen.cs", SearchOption.AllDirectories))
+                    {
+                        File.Delete(file);
+                    }
                 }
             }
 
             for (var i = 0; i < config.Tasks.Length; i++)
             {
                 var i1 = i;
-                if (TestMode)
+                if (args.Contains("no-parallel"))
                 {
-                    RunTask(config.Tasks[i1]);
+                    RunTask(config.Tasks[i1], args);
                 }
                 else
                 {
-                    tasks[i] = Task.Run(() => RunTaskGuarded(config.Tasks[i1]));
+                    tasks[i] = Task.Run(() => RunTaskGuarded(config.Tasks[i1], args));
                 }
             }
 
-            if (!TestMode)
+            if (!args.Contains("no-parallel"))
             {
                 Task.WaitAll(tasks);
             }
         }
 
-        public static void RunTaskGuarded(BindTask task)
+        public static void RunTaskGuarded(BindTask task, IReadOnlyCollection<string> args)
         {
             try
             {
-                RunTask(task);
+                RunTask(task, args);
             }
             catch (Exception ex)
             {
@@ -70,7 +72,7 @@ namespace Silk.NET.BuildTools
             }
         }
 
-        public static void RunTask(BindTask task)
+        public static void RunTask(BindTask task, IReadOnlyCollection<string> args)
         {
             Stopwatch sw = null;
             if (!(Program.ConsoleWriter.Instance is null))
@@ -102,110 +104,145 @@ namespace Silk.NET.BuildTools
                 }
             }
             
-            Profile profile;
-            if (ShouldConvert(task.Controls))
+            Profile profile = null;
+            var cacheHit = !ShouldConvert(task.Controls, args);
+            if (!cacheHit)
             {
-                Console.WriteLine("Profile conversion started!");
-                var tsb4 = sw?.Elapsed.TotalSeconds;
-                var profiles = new List<Profile>();
-                if (task.Mode == ConverterMode.ConvertConstruct)
+                try
                 {
-                    foreach (var src in task.Sources)
+                    Console.WriteLine("Profile conversion started!");
+                    var tsb4 = sw?.Elapsed.TotalSeconds;
+                    var profiles = new List<Profile>();
+                    if (task.Mode == ConverterMode.ConvertConstruct)
                     {
-                        var rawProfiles = ProfileConverter.ReadProfiles
-                        (
-                            task.ConverterOpts.Reader.ToLower() switch
-                            {
-                                "gl" => new OpenGLReader(),
-                                "cl" => new OpenCLReader(),
-                                "vk" => new VulkanReader(),
-                                _ => throw new ArgumentException("Couldn't find a reader with that name")
-                            }, task.ConverterOpts.Constructor.ToLower() switch
-                            {
-                                "gl" => new OpenGLConstructor(),
-                                "cl" => new OpenCLConstructor(),
-                                "vk" => new VulkanConstructor(),
-                                _ => throw new ArgumentException("Couldn't find a reader with that name")
-                            },
-                            OpenPath(src),
-                            task
-                        ).ToList();
-
-                        Console.WriteLine("Raw profile parsing complete, cloning in memory prior to baking...");
-                        profiles.AddRange
-                        (
-                            // BUG this is an awful fix for a weird bug, but if we don't do this everything falls apart.
-                            // feel free to remove the serialize-deserialize and try for yourself would welcome a fix ;)
-                            JsonConvert.DeserializeObject<Profile[]>
-                            (
-                                JsonConvert.SerializeObject
+                        foreach (var src in task.Sources)
+                        {
+                            var rawProfiles = ProfileConverter.ReadProfiles
                                 (
-                                    rawProfiles
+                                    task.ConverterOpts.Reader.ToLower() switch
+                                    {
+                                        "gl" => new OpenGLReader(),
+                                        "cl" => new OpenCLReader(),
+                                        "vk" => new VulkanReader(),
+                                        _ => throw new ArgumentException("Couldn't find a reader with that name")
+                                    }, task.ConverterOpts.Constructor.ToLower() switch
+                                    {
+                                        "gl" => new OpenGLConstructor(),
+                                        "cl" => new OpenCLConstructor(),
+                                        "vk" => new VulkanConstructor(),
+                                        _ => throw new ArgumentException("Couldn't find a reader with that name")
+                                    },
+                                    OpenPath(src),
+                                    task
                                 )
-                            )
-                        );
-                        
-                        Console.WriteLine("Profiles are ready.");
+                                .ToList();
+
+                            Console.WriteLine("Raw profile parsing complete, cloning in memory prior to baking...");
+                            profiles.AddRange
+                            (
+                                // BUG this is an awful fix for a weird bug, but if we don't do this everything falls apart.
+                                // feel free to remove the serialize-deserialize and try for yourself would welcome a fix ;)
+                                JsonConvert.DeserializeObject<Profile[]>
+                                (
+                                    JsonConvert.SerializeObject
+                                    (
+                                        rawProfiles
+                                    )
+                                )
+                            );
+
+                            Console.WriteLine("Profiles are ready.");
+                        }
+                    }
+                    else if (task.Mode == ConverterMode.Clang)
+                    {
+                        foreach (var src in task.Sources)
+                        {
+                            profiles.Add(Clang.GenerateProfile(Path.GetFileName(src), OpenPath(src), task));
+                        }
+                    }
+
+                    profile = ProfileBakery.Bake
+                        (task.Name, profiles.Where(x => task.BakeryOpts.Include.Contains(x.Name)).ToList());
+
+                    var tsaf = sw?.Elapsed.TotalSeconds - tsb4;
+                    var tsafTxt = sw is null ? null : $", took {tsaf} second(s)";
+                    Console.WriteLine($"Conversion complete{tsafTxt}.");
+
+                    try
+                    {
+                        if (!string.IsNullOrWhiteSpace(task.CacheKey) && !string.IsNullOrWhiteSpace(task.CacheFolder))
+                        {
+                            if (!Directory.Exists(task.CacheFolder))
+                            {
+                                Directory.CreateDirectory(task.CacheFolder);
+                            }
+
+                            using var fileStream = File.OpenWrite
+                                (Path.Combine(task.CacheFolder, task.CacheKey + ".json.gz"));
+                            using var gzStream = new GZipStream(fileStream, CompressionLevel.Optimal);
+                            gzStream.Write(Encoding.UTF8.GetBytes(JsonConvert.SerializeObject(profile)));
+                            gzStream.Flush();
+                            fileStream.Flush();
+                            Console.WriteLine("Written to cache for future use.");
+                        }
+                    }
+                    catch (Exception ex)
+                    {
+                        Console.WriteLine("Couldn't write to cache, next cache hit may be outdated.");
+                        Console.WriteLine(ex);
                     }
                 }
-                else if (task.Mode == ConverterMode.Clang)
+                catch (Exception ex)
                 {
-                    foreach (var src in task.Sources)
-                    {
-                        profiles.Add(Clang.GenerateProfile(Path.GetFileName(src), OpenPath(src), task));
-                    }
-                }
-
-                profile = ProfileBakery.Bake
-                    (task.Name, profiles.Where(x => task.BakeryOpts.Include.Contains(x.Name)).ToList());
-
-                var tsaf = sw?.Elapsed.TotalSeconds - tsb4;
-                var tsafTxt = sw is null ? null : $", took {tsaf} second(s)";
-                Console.WriteLine($"Conversion complete{tsafTxt}.");
-
-                if (!string.IsNullOrWhiteSpace(task.CacheKey) && !string.IsNullOrWhiteSpace(task.CacheFolder))
-                {
-                    if (!Directory.Exists(task.CacheFolder))
-                    {
-                        Directory.CreateDirectory(task.CacheFolder);
-                    }
-                
-                    using var fileStream = File.OpenWrite(Path.Combine(task.CacheFolder, task.CacheKey + ".json.gz"));
-                    using var gzStream = new GZipStream(fileStream, CompressionLevel.Optimal);
-                    gzStream.Write(Encoding.UTF8.GetBytes(JsonConvert.SerializeObject(profile)));
-                    gzStream.Flush();
-                    fileStream.Flush();
-                    Console.WriteLine("Written to cache for future use.");
+                    Console.WriteLine("Profile conversion failed.");
+                    Console.WriteLine(ex);
+                    Console.WriteLine("Attempting to continue with a cache hit...");
+                    cacheHit = true;
                 }
             }
-            else if (!string.IsNullOrWhiteSpace(task.CacheKey) && !string.IsNullOrWhiteSpace(task.CacheFolder))
+
+            if (cacheHit)
             {
-                Console.WriteLine("Cache hit!");
-                var tsb4 = sw?.Elapsed.TotalSeconds;
-                var file = Path.Combine(task.CacheFolder, task.CacheKey + ".json.gz");
-                if (!File.Exists(file))
+                if (!string.IsNullOrWhiteSpace(task.CacheKey) && !string.IsNullOrWhiteSpace(task.CacheFolder))
+                {
+                    Console.WriteLine("Cache hit!");
+                    var tsb4 = sw?.Elapsed.TotalSeconds;
+                    var file = Path.Combine(task.CacheFolder, task.CacheKey + ".json.gz");
+                    if (!File.Exists(file))
+                    {
+                        throw new InvalidOperationException
+                        (
+                            "Couldn't find a cached profile to fallback on" +
+                            "(conversion was skipped as per the control variables)"
+                        );
+                    }
+
+                    using var memoryStream = new MemoryStream();
+                    using var fileStream = File.OpenRead(file);
+                    using var gzStream = new GZipStream(fileStream, CompressionMode.Decompress);
+                    gzStream.CopyTo(memoryStream);
+                    profile = JsonConvert.DeserializeObject<Profile>(Encoding.UTF8.GetString(memoryStream.ToArray()));
+                    var tsaf = tsb4 - sw?.Elapsed.TotalSeconds;
+                    var tsafTxt = sw is null ? null : $", took {tsaf} second(s)";
+                    Console.WriteLine($"Cached profile loaded{tsafTxt}.");
+                }
+                else
                 {
                     throw new InvalidOperationException
                     (
                         "Couldn't find a cached profile to fallback on" +
-                        "(conversion was skipped as per the control variables)"
+                        "(conversion was skipped per the control variables)"
                     );
                 }
-                using var memoryStream = new MemoryStream();
-                using var fileStream = File.OpenRead(file);
-                using var gzStream = new GZipStream(fileStream, CompressionMode.Decompress);
-                gzStream.CopyTo(memoryStream);
-                profile = JsonConvert.DeserializeObject<Profile>(Encoding.UTF8.GetString(memoryStream.ToArray()));
-                var tsaf = tsb4 - sw?.Elapsed.TotalSeconds;
-                var tsafTxt = sw is null ? null : $", took {tsaf} second(s)";
-                Console.WriteLine($"Cached profile loaded{tsafTxt}.");
             }
-            else
+
+            if (profile is null)
             {
-                throw new InvalidOperationException
+                throw new NullReferenceException
                 (
-                    "Couldn't find a cached profile to fallback on" +
-                    "(conversion was skipped as per the control variables)"
+                    "Profile is null. This shouldn't be possible, and indicates a fatal error in both converting " +
+                    "and cache hitting."
                 );
             }
 
@@ -219,8 +256,18 @@ namespace Silk.NET.BuildTools
                     new KeyValuePair<string, TimeSpan>(task.Name, sw.Elapsed);
             }
             
-            static bool ShouldConvert(string[] controls)
+            static bool ShouldConvert(string[] controls, IReadOnlyCollection<string> args)
             {
+                if (args.Contains("no-cache-hit"))
+                {
+                    return true;
+                }
+
+                if (args.Contains("always-cache-hit"))
+                {
+                    return false;
+                }
+                
                 if (controls.Any
                     (y => y.ToLower() == "convert-windows-only") && !RuntimeInformation.IsOSPlatform
                     (OSPlatform.Windows))
